==> ie_script/2.Split.ipynb <==
### In order to facilitate processing, split data

import time

import pickle
import pandas as pd
import spacy

from tqdm import tqdm

import numpy as np

### Read file

in_file  = "./cleaned_data.sav"
df = pickle.load(open(in_file, 'rb'))

len(df) # Total number of scientific articles

df_split = np.array_split(df, 10)

### Split

for i in tqdm(range(len(df_split))):
    abstract = []
    body_text = []
    for article in df_split[i]['abstract']:
        abstract.append(article)
    for article in df_split[i]['body_text']:
        body_text.append(article)
    pickle.dump(abstract, open('./data_split/abstract_'+str(i)+'.sav', 'wb'))
    pickle.dump(abstract, open('./data_split/body_'+str(i)+'.sav', 'wb'))
