==> ie_script/4.Biobert.ipynb <==
import time

import pickle
import pandas as pd

from tqdm import tqdm
import numpy as np

file cleaned_data.sav from step 1.Pre.ipynb

in_file  = "./cleaned_data.sav"
df = pd.read_pickle(open(in_file, 'rb'))

df_split = np.array_split(df, 10)
for i in tqdm(range(len(df_split))):
    body_text = []
    for article in df_split[i]['body_text']:
        body_text.append(article)
    pickle.dump(body_text, open('./body_text/body_text_'+str(i)+'.sav', 'wb'))

sav to json

import spacy
import json

nlp = spacy.load('en_core_web_lg')
nlp.max_length = 2000000

def to_sent(text):
    doc=nlp(text)
    result=[]
    for sent in list(doc.sents):
        sent=str(sent)
        sent = sent.rstrip() #remove end space
        result.append(sent)
    return result

def to_json(df, save_file_name):
    results = []
    
    for text in tqdm(df):
        results += to_sent(text)
    
    json_file = save_file_name + '.json'
    with open(json_file,'w') as f:
        json.dump(results, f)

# abstract
abstract = df['abstract']

to_json(abstract, 'abstract')

# body_text
filename = '0' # 1 2 3 4 ... 9
df = pd.read_pickle(open('./body_text/body_text_'+filename+'.sav', 'rb'))
to_json(df, './body_text/body_text_'+filename+'.json')

minie

# in file Minie_Bash.ipynb

filter and label sentence for biobert

pronounsList = ["another", "anybody", "anyone", "anything", "each", "either", "enough", "everybody", "everyone",
                "everything", "little", "much", "neither", "nobody", "no one", "nothing", "one", "other", "somebody",
                "something", "both", "few", "fewer", "many", "others", "several", "all", "any", "more", "most",
                "someone", "none", "some", "such", "I", "me", "my", "mine", "you", "your", "yours", "yourself",
                "yourselves", "he", "him", "his", "himself", "she", "her", "hers", "herself", "myself", "it", "its", 
                "itself", "we", "us", "our", "ours", "ourselves", "they", "them", "their", "theirs", "themselves"]

def label(csv_file, name):
    csv_data = pd.read_csv(csv_file,encoding='utf-8')
    csv_df = pd.DataFrame(csv_data)
    
    # remove QUANT_
    df = csv_df.dropna()[~csv_df.dropna()['subject'].str.contains('QUANT_')]
    df = df.dropna()[~df.dropna()['object'].str.contains('QUANT_')]
    
    # remove single pronoun
    df = df.dropna()[~df.dropna()['subject'].str.lower().isin(pronounsList)]
    df = df.dropna()[~df.dropna()['object'].str.lower().isin(pronounsList)]
    
    # remove \n
    df['sentence'] = df['sentence'].replace(r'\s+|\\n', ' ', regex=True)
    
    # label sent
    def label1(sub,obj,sent):
        sub_location = sent.find(sub)
        if(sub_location == -1):
            return sent
        new_sub = '@' + sub.upper() + '#'
        sent = sent[:sub_location] + new_sub + sent[sub_location+len(sub):]
        obj_location = sent.find(obj)
        if(obj_location == -1):
            return sent
        new_obj = '@' + obj.upper() + '#'
        sent = sent[:obj_location] + new_obj + sent[obj_location+len(obj):]
        return sent
    df['sent'] = [ label1(a,b,c) for a,b,c in zip(df['subject'],df['object'],df['sentence']) ]
    
    # set label
    def label2(a):
        if(a[0].count('@') >= 2 and a[0].count('#') >= 2):
            return 1
        else:
            return 0
    df['label'] = [ label2(a) for a in zip(df['sent']) ]
    
    # save all
    df.insert(0, 'index', range(len(df)), allow_duplicates=False)
    df.to_csv('./for_biobert/'+ 'labeled_'+csv_file , index=False)
    
    # save for test.tsv
    df=df.drop('subject',axis=1)
    df=df.drop('relation',axis=1)
    df=df.drop('object',axis=1)
    df=df.drop('sentence',axis=1)
    df.columns = ['index','sentence','label']
    
    df.to_csv('./for_biobert/'+ name + '_test.tsv', sep="\t" , index=False)
    

label('body_text_10.json.csv', 'body_text_10')

run biobert

# in file ./biobert/Biobert_Bash.ipynb

# read tsv
tsv = pd.read_csv('./for_biobert/body_text_1_test.tsv', sep='\t')

# read result_csv
result_csv = pd.read_csv('./for_biobert/body_text_1_test_result差.csv',encoding='utf-8')
result_csv3 = pd.read_csv('./for_biobert/body_text_3_test_result差.csv',encoding='utf-8')
result_csv4 = pd.read_csv('./for_biobert/body_text_4_test_result差.csv',encoding='utf-8')

# read labeled_csv
labeled_csv = pd.read_csv('./for_biobert/labeled_body_text_1.json.csv',encoding='utf-8')
